move commands into their own files
