+++ src/storage/txn/commands/check_txn_status.rs
@@ -1,0 +1,43 @@
+use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::txn::latch::{self, Latches};
+use crate::storage::{Context, TimeStamp, TxnStatus};
+use crate::{command, ctx, gen_lock, tag, ts, write_bytes};
+use std::fmt::{self, Debug, Display, Formatter};
+use std::iter;
+use txn_types::Key;
+
+command! {
+    /// Check the status of a transaction. This is usually invoked by a transaction that meets
+    /// another transaction's lock. If the primary lock is expired, it will rollback the primary
+    /// lock. If the primary lock exists but is not expired, it may update the transaction's
+    /// `min_commit_ts`. Returns a [`TxnStatus`](TxnStatus) to represent the status.
+    ///
+    /// This is invoked on a transaction's primary lock. The lock may be generated by either
+    /// [`AcquirePessimisticLock`](Command::AcquirePessimisticLock) or
+    /// [`Prewrite`](Command::Prewrite).
+    CheckTxnStatus:
+        cmd_ty => TxnStatus,
+        display => "kv::command::check_txn_status {} @ {} curr({}, {}) | {:?}", (primary_key, lock_ts, caller_start_ts, current_ts, ctx),
+        content => {
+            /// The primary key of the transaction.
+            primary_key: Key,
+            /// The lock's ts, namely the transaction's start_ts.
+            lock_ts: TimeStamp,
+            /// The start_ts of the transaction that invokes this command.
+            caller_start_ts: TimeStamp,
+            /// The approximate current_ts when the command is invoked.
+            current_ts: TimeStamp,
+            /// Specifies the behavior when neither commit/rollback record nor lock is found. If true,
+            /// rollbacks that transaction; otherwise returns an error.
+            rollback_if_not_exist: bool,
+        }
+}
+
+impl CommandExt for CheckTxnStatus {
+    ctx!();
+    tag!(check_txn_status);
+    ts!(lock_ts);
+    write_bytes!(primary_key);
+    gen_lock!(primary_key);
+}
