+++ src/storage/txn/commands/mod.rs
@@ -1,18 +1,33 @@
 // Copyright 2019 TiKV Project Authors. Licensed under Apache-2.0.
+pub mod acquire_pessimistic_lock;
+pub mod check_txn_status;
+pub mod cleanup;
+pub mod commit;
+pub mod mvcc_by_key;
+pub mod mvcc_by_start_ts;
+pub mod pause;
+pub mod pessimistic_rollback;
+pub mod prewrite;
+pub mod prewrite_pessimistic;
+pub mod resolve_lock;
+pub mod resolve_lock_lite;
+pub mod rollback;
+pub mod scan_lock;
+pub mod txn_heart_beat;
 
 use std::fmt::{self, Debug, Display, Formatter};
 use std::iter::{self, FromIterator};
 use std::marker::PhantomData;
 
 use kvproto::kvrpcpb::*;
-use tikv_util::collections::HashMap;
-use txn_types::{Key, Lock, Mutation, TimeStamp};
+use txn_types::{Key, TimeStamp};
 
 use crate::storage::lock_manager::WaitTimeout;
-use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+use crate::storage::metrics;
 use crate::storage::txn::latch::{self, Latches};
 use crate::storage::types::{MvccInfo, PessimisticLockRes, StorageCallbackType, TxnStatus};
 use crate::storage::Result;
+use tikv_util::collections::HashMap;
 
 /// Store Transaction scheduler commands.
 ///
@@ -23,21 +38,21 @@ use crate::storage::Result;
 /// [`Storage::prewrite`](Storage::prewrite) trait and are executed asynchronously.
 // Logic related to these can be found in the `src/storage/txn/proccess.rs::process_write_impl` function.
 pub enum Command {
-    Prewrite(Prewrite),
-    PrewritePessimistic(PrewritePessimistic),
-    AcquirePessimisticLock(AcquirePessimisticLock),
-    Commit(Commit),
-    Cleanup(Cleanup),
-    Rollback(Rollback),
-    PessimisticRollback(PessimisticRollback),
-    TxnHeartBeat(TxnHeartBeat),
-    CheckTxnStatus(CheckTxnStatus),
-    ScanLock(ScanLock),
-    ResolveLock(ResolveLock),
-    ResolveLockLite(ResolveLockLite),
-    Pause(Pause),
-    MvccByKey(MvccByKey),
-    MvccByStartTs(MvccByStartTs),
+    Prewrite(prewrite::Prewrite),
+    PrewritePessimistic(prewrite_pessimistic::PrewritePessimistic),
+    AcquirePessimisticLock(acquire_pessimistic_lock::AcquirePessimisticLock),
+    Commit(commit::Commit),
+    Cleanup(cleanup::Cleanup),
+    Rollback(rollback::Rollback),
+    PessimisticRollback(pessimistic_rollback::PessimisticRollback),
+    TxnHeartBeat(txn_heart_beat::TxnHeartBeat),
+    CheckTxnStatus(check_txn_status::CheckTxnStatus),
+    ResolveLock(resolve_lock::ResolveLock),
+    ResolveLockLite(resolve_lock_lite::ResolveLockLite),
+    ScanLock(scan_lock::ScanLock),
+    Pause(pause::Pause),
+    MvccByKey(mvcc_by_key::MvccByKey),
+    MvccByStartTs(mvcc_by_start_ts::MvccByStartTs),
 }
 
 pub struct TypedCommand<T> {
@@ -64,7 +79,7 @@ impl From<PrewriteRequest> for TypedCommand<Vec<Result<()>>> {
     fn from(mut req: PrewriteRequest) -> Self {
         let for_update_ts = req.get_for_update_ts();
         if for_update_ts == 0 {
-            Prewrite::new(
+            prewrite::Prewrite::new(
                 req.take_mutations().into_iter().map(Into::into).collect(),
                 req.take_primary_lock(),
                 req.get_start_version().into(),
@@ -82,7 +97,7 @@ impl From<PrewriteRequest> for TypedCommand<Vec<Result<()>>> {
                 .map(Into::into)
                 .zip(is_pessimistic_lock.into_iter())
                 .collect();
-            PrewritePessimistic::new(
+            prewrite_pessimistic::PrewritePessimistic::new(
                 mutations,
                 req.take_primary_lock(),
                 req.get_start_version().into(),
@@ -110,7 +125,7 @@ impl From<PessimisticLockRequest> for TypedCommand<Result<PessimisticLockRes>> {
             })
             .collect();
 
-        AcquirePessimisticLock::new(
+        acquire_pessimistic_lock::AcquirePessimisticLock::new(
             keys,
             req.take_primary_lock(),
             req.get_start_version().into(),
@@ -129,7 +144,7 @@ impl From<CommitRequest> for TypedCommand<TxnStatus> {
     fn from(mut req: CommitRequest) -> Self {
         let keys = req.get_keys().iter().map(|x| Key::from_raw(x)).collect();
 
-        Commit::new(
+        commit::Commit::new(
             keys,
             req.get_start_version().into(),
             req.get_commit_version().into(),
@@ -140,7 +155,7 @@ impl From<CommitRequest> for TypedCommand<TxnStatus> {
 
 impl From<CleanupRequest> for TypedCommand<()> {
     fn from(mut req: CleanupRequest) -> Self {
-        Cleanup::new(
+        cleanup::Cleanup::new(
             Key::from_raw(req.get_key()),
             req.get_start_version().into(),
             req.get_current_ts().into(),
@@ -152,7 +167,7 @@ impl From<CleanupRequest> for TypedCommand<()> {
 impl From<BatchRollbackRequest> for TypedCommand<()> {
     fn from(mut req: BatchRollbackRequest) -> Self {
         let keys = req.get_keys().iter().map(|x| Key::from_raw(x)).collect();
-        Rollback::new(keys, req.get_start_version().into(), req.take_context())
+        rollback::Rollback::new(keys, req.get_start_version().into(), req.take_context())
     }
 }
 
@@ -160,7 +175,7 @@ impl From<PessimisticRollbackRequest> for TypedCommand<Vec<Result<()>>> {
     fn from(mut req: PessimisticRollbackRequest) -> Self {
         let keys = req.get_keys().iter().map(|x| Key::from_raw(x)).collect();
 
-        PessimisticRollback::new(
+        pessimistic_rollback::PessimisticRollback::new(
             keys,
             req.get_start_version().into(),
             req.get_for_update_ts().into(),
@@ -171,7 +186,7 @@ impl From<PessimisticRollbackRequest> for TypedCommand<Vec<Result<()>>> {
 
 impl From<TxnHeartBeatRequest> for TypedCommand<TxnStatus> {
     fn from(mut req: TxnHeartBeatRequest) -> Self {
-        TxnHeartBeat::new(
+        txn_heart_beat::TxnHeartBeat::new(
             Key::from_raw(req.get_primary_lock()),
             req.get_start_version().into(),
             req.get_advise_lock_ttl(),
@@ -182,7 +197,7 @@ impl From<TxnHeartBeatRequest> for TypedCommand<TxnStatus> {
 
 impl From<CheckTxnStatusRequest> for TypedCommand<TxnStatus> {
     fn from(mut req: CheckTxnStatusRequest) -> Self {
-        CheckTxnStatus::new(
+        check_txn_status::CheckTxnStatus::new(
             Key::from_raw(req.get_primary_key()),
             req.get_lock_ts().into(),
             req.get_caller_start_ts().into(),
@@ -201,7 +216,7 @@ impl From<ScanLockRequest> for TypedCommand<Vec<LockInfo>> {
             Some(Key::from_raw(req.get_start_key()))
         };
 
-        ScanLock::new(
+        scan_lock::ScanLock::new(
             req.get_max_version().into(),
             start_key,
             req.get_limit() as usize,
@@ -231,25 +246,30 @@ impl From<ResolveLockRequest> for TypedCommand<()> {
         };
 
         if resolve_keys.is_empty() {
-            ResolveLock::new(txn_status, None, vec![], req.take_context())
+            resolve_lock::ResolveLock::new(txn_status, None, vec![], req.take_context())
         } else {
             let start_ts: TimeStamp = req.get_start_version().into();
             assert!(!start_ts.is_zero());
             let commit_ts = req.get_commit_version().into();
-            ResolveLockLite::new(start_ts, commit_ts, resolve_keys, req.take_context())
+            resolve_lock_lite::ResolveLockLite::new(
+                start_ts,
+                commit_ts,
+                resolve_keys,
+                req.take_context(),
+            )
         }
     }
 }
 
 impl From<MvccGetByKeyRequest> for TypedCommand<MvccInfo> {
     fn from(mut req: MvccGetByKeyRequest) -> Self {
-        MvccByKey::new(Key::from_raw(req.get_key()), req.take_context())
+        mvcc_by_key::MvccByKey::new(Key::from_raw(req.get_key()), req.take_context())
     }
 }
 
 impl From<MvccGetByStartTsRequest> for TypedCommand<Option<(Key, MvccInfo)>> {
     fn from(mut req: MvccGetByStartTsRequest) -> Self {
-        MvccByStartTs::new(req.get_start_ts().into(), req.take_context())
+        mvcc_by_start_ts::MvccByStartTs::new(req.get_start_ts().into(), req.take_context())
     }
 }
 
@@ -287,6 +307,7 @@ pub trait CommandExt: Display {
     fn gen_lock(&self, _latches: &Latches) -> latch::Lock;
 }
 
+#[macro_export]
 macro_rules! ctx {
     () => {
         fn get_ctx(&self) -> &Context {
@@ -298,6 +319,7 @@ macro_rules! ctx {
     };
 }
 
+#[macro_export]
 macro_rules! command {
     (
         $(#[$outer_doc: meta])*
@@ -346,6 +368,7 @@ macro_rules! command {
     }
 }
 
+#[macro_export]
 macro_rules! ts {
     ($ts:ident) => {
         fn ts(&self) -> TimeStamp {
@@ -354,6 +377,7 @@ macro_rules! ts {
     };
 }
 
+#[macro_export]
 macro_rules! tag {
     ($tag:ident) => {
         fn tag(&self) -> metrics::CommandKind {
@@ -366,6 +390,7 @@ macro_rules! tag {
     };
 }
 
+#[macro_export]
 macro_rules! write_bytes {
     ($field: ident) => {
         fn write_bytes(&self) -> usize {
@@ -379,6 +404,7 @@ macro_rules! write_bytes {
     };
 }
 
+#[macro_export]
 macro_rules! gen_lock {
     (empty) => {
         fn gen_lock(&self, _latches: &Latches) -> latch::Lock {
@@ -404,540 +430,13 @@ macro_rules! gen_lock {
     };
 }
 
+#[macro_export]
 macro_rules! command_method {
     ($name:ident, $return_ty: ty, $value: expr) => {
         fn $name(&self) -> $return_ty {
             $value
         }
     };
-}
-
-command! {
-    /// The prewrite phase of a transaction. The first phase of 2PC.
-    ///
-    /// This prepares the system to commit the transaction. Later a [`Commit`](Command::Commit)
-    /// or a [`Rollback`](Command::Rollback) should follow.
-    Prewrite:
-        cmd_ty => Vec<Result<()>>,
-        display => "kv::command::prewrite mutations({}) @ {} | {:?}", (mutations.len, start_ts, ctx),
-        content => {
-            /// The set of mutations to apply.
-            mutations: Vec<Mutation>,
-            /// The primary lock. Secondary locks (from `mutations`) will refer to the primary lock.
-            primary: Vec<u8>,
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-            lock_ttl: u64,
-            skip_constraint_check: bool,
-            /// How many keys this transaction involved.
-            txn_size: u64,
-            min_commit_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for Prewrite {
-    ctx!();
-    tag!(prewrite);
-    ts!(start_ts);
-
-    fn write_bytes(&self) -> usize {
-        let mut bytes = 0;
-        for m in &self.mutations {
-            match *m {
-                Mutation::Put((ref key, ref value)) | Mutation::Insert((ref key, ref value)) => {
-                    bytes += key.as_encoded().len();
-                    bytes += value.len();
-                }
-                Mutation::Delete(ref key) | Mutation::Lock(ref key) => {
-                    bytes += key.as_encoded().len();
-                }
-                Mutation::CheckNotExists(_) => (),
-            }
-        }
-        bytes
-    }
-
-    gen_lock!(mutations: multiple(|x| x.key()));
-}
-
-impl Prewrite {
-    #[cfg(test)]
-    pub fn with_defaults(
-        mutations: Vec<Mutation>,
-        primary: Vec<u8>,
-        start_ts: TimeStamp,
-    ) -> TypedCommand<Vec<Result<()>>> {
-        Prewrite::new(
-            mutations,
-            primary,
-            start_ts,
-            0,
-            false,
-            0,
-            TimeStamp::default(),
-            Context::default(),
-        )
-    }
-
-    #[cfg(test)]
-    pub fn with_lock_ttl(
-        mutations: Vec<Mutation>,
-        primary: Vec<u8>,
-        start_ts: TimeStamp,
-        lock_ttl: u64,
-    ) -> TypedCommand<Vec<Result<()>>> {
-        Prewrite::new(
-            mutations,
-            primary,
-            start_ts,
-            lock_ttl,
-            false,
-            0,
-            TimeStamp::default(),
-            Context::default(),
-        )
-    }
-
-    pub fn with_context(
-        mutations: Vec<Mutation>,
-        primary: Vec<u8>,
-        start_ts: TimeStamp,
-        ctx: Context,
-    ) -> TypedCommand<Vec<Result<()>>> {
-        Prewrite::new(
-            mutations,
-            primary,
-            start_ts,
-            0,
-            false,
-            0,
-            TimeStamp::default(),
-            ctx,
-        )
-    }
-}
-
-command! {
-    /// The prewrite phase of a transaction using pessimistic locking. The first phase of 2PC.
-    ///
-    /// This prepares the system to commit the transaction. Later a [`Commit`](Command::Commit)
-    /// or a [`Rollback`](Command::Rollback) should follow.
-    PrewritePessimistic:
-        cmd_ty => Vec<Result<()>>,
-        display => "kv::command::prewrite_pessimistic mutations({}) @ {} | {:?}", (mutations.len, start_ts, ctx),
-        content => {
-            /// The set of mutations to apply; the bool = is pessimistic lock.
-            mutations: Vec<(Mutation, bool)>,
-            /// The primary lock. Secondary locks (from `mutations`) will refer to the primary lock.
-            primary: Vec<u8>,
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-            lock_ttl: u64,
-            for_update_ts: TimeStamp,
-            /// How many keys this transaction involved.
-            txn_size: u64,
-            min_commit_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for PrewritePessimistic {
-    ctx!();
-    tag!(prewrite);
-    ts!(start_ts);
-    command_method!(requires_pessimistic_txn, bool, true);
-
-    fn write_bytes(&self) -> usize {
-        let mut bytes = 0;
-        for (m, _) in &self.mutations {
-            match *m {
-                Mutation::Put((ref key, ref value)) | Mutation::Insert((ref key, ref value)) => {
-                    bytes += key.as_encoded().len();
-                    bytes += value.len();
-                }
-                Mutation::Delete(ref key) | Mutation::Lock(ref key) => {
-                    bytes += key.as_encoded().len();
-                }
-                Mutation::CheckNotExists(_) => (),
-            }
-        }
-        bytes
-    }
-
-    gen_lock!(mutations: multiple(|(x, _)| x.key()));
-}
-
-command! {
-    /// Acquire a Pessimistic lock on the keys.
-    ///
-    /// This can be rolled back with a [`PessimisticRollback`](Command::PessimisticRollback) command.
-    AcquirePessimisticLock:
-        cmd_ty => Result<PessimisticLockRes>,
-        display => "kv::command::acquirepessimisticlock keys({}) @ {} {} | {:?}", (keys.len, start_ts, for_update_ts, ctx),
-        content => {
-            /// The set of keys to lock.
-            keys: Vec<(Key, bool)>,
-            /// The primary lock. Secondary locks (from `keys`) will refer to the primary lock.
-            primary: Vec<u8>,
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-            lock_ttl: u64,
-            is_first_lock: bool,
-            for_update_ts: TimeStamp,
-            /// Time to wait for lock released in milliseconds when encountering locks.
-            wait_timeout: Option<WaitTimeout>,
-            /// If it is true, TiKV will return values of the keys if no error, so TiDB can cache the values for
-            /// later read in the same transaction.
-            return_values: bool,
-            min_commit_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for AcquirePessimisticLock {
-    ctx!();
-    tag!(acquire_pessimistic_lock);
-    ts!(start_ts);
-    command_method!(requires_pessimistic_txn, bool, true);
-    command_method!(can_be_pipelined, bool, true);
-
-    fn write_bytes(&self) -> usize {
-        self.keys
-            .iter()
-            .map(|(key, _)| key.as_encoded().len())
-            .sum()
-    }
-
-    gen_lock!(keys: multiple(|x| &x.0));
-}
-
-command! {
-    /// Commit the transaction that started at `lock_ts`.
-    ///
-    /// This should be following a [`Prewrite`](Command::Prewrite).
-    Commit:
-        cmd_ty => TxnStatus,
-        display => "kv::command::commit {} {} -> {} | {:?}", (keys.len, lock_ts, commit_ts, ctx),
-        content => {
-            /// The keys affected.
-            keys: Vec<Key>,
-            /// The lock timestamp.
-            lock_ts: TimeStamp,
-            /// The commit timestamp.
-            commit_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for Commit {
-    ctx!();
-    tag!(commit);
-    ts!(commit_ts);
-    write_bytes!(keys: multiple);
-    gen_lock!(keys: multiple);
-}
-
-command! {
-    /// Rollback mutations on a single key.
-    ///
-    /// This should be following a [`Prewrite`](Command::Prewrite) on the given key.
-    Cleanup:
-        cmd_ty => (),
-        display => "kv::command::cleanup {} @ {} | {:?}", (key, start_ts, ctx),
-        content => {
-            key: Key,
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-            /// The approximate current ts when cleanup request is invoked, which is used to check the
-            /// lock's TTL. 0 means do not check TTL.
-            current_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for Cleanup {
-    ctx!();
-    tag!(cleanup);
-    ts!(start_ts);
-    write_bytes!(key);
-    gen_lock!(key);
-}
-
-command! {
-    /// Rollback from the transaction that was started at `start_ts`.
-    ///
-    /// This should be following a [`Prewrite`](Command::Prewrite) on the given key.
-    Rollback:
-        cmd_ty => (),
-        display => "kv::command::rollback keys({}) @ {} | {:?}", (keys.len, start_ts, ctx),
-        content => {
-            keys: Vec<Key>,
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for Rollback {
-    ctx!();
-    tag!(rollback);
-    ts!(start_ts);
-    write_bytes!(keys: multiple);
-    gen_lock!(keys: multiple);
-}
-
-command! {
-    /// Rollback pessimistic locks identified by `start_ts` and `for_update_ts`.
-    ///
-    /// This can roll back an [`AcquirePessimisticLock`](Command::AcquirePessimisticLock) command.
-    PessimisticRollback:
-        cmd_ty => Vec<Result<()>>,
-        display => "kv::command::pessimistic_rollback keys({}) @ {} {} | {:?}", (keys.len, start_ts, for_update_ts, ctx),
-        content => {
-            /// The keys to be rolled back.
-            keys: Vec<Key>,
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-            for_update_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for PessimisticRollback {
-    ctx!();
-    tag!(pessimistic_rollback);
-    ts!(start_ts);
-    command_method!(requires_pessimistic_txn, bool, true);
-    write_bytes!(keys: multiple);
-    gen_lock!(keys: multiple);
-}
-
-command! {
-    /// Heart beat of a transaction. It enlarges the primary lock's TTL.
-    ///
-    /// This is invoked on a transaction's primary lock. The lock may be generated by either
-    /// [`AcquirePessimisticLock`](Command::AcquirePessimisticLock) or
-    /// [`Prewrite`](Command::Prewrite).
-    TxnHeartBeat:
-        cmd_ty => TxnStatus,
-        display => "kv::command::txn_heart_beat {} @ {} ttl {} | {:?}", (primary_key, start_ts, advise_ttl, ctx),
-        content => {
-            /// The primary key of the transaction.
-            primary_key: Key,
-            /// The transaction's start_ts.
-            start_ts: TimeStamp,
-            /// The new TTL that will be used to update the lock's TTL. If the lock's TTL is already
-            /// greater than `advise_ttl`, nothing will happen.
-            advise_ttl: u64,
-        }
-}
-
-impl CommandExt for TxnHeartBeat {
-    ctx!();
-    tag!(txn_heart_beat);
-    ts!(start_ts);
-    write_bytes!(primary_key);
-    gen_lock!(primary_key);
-}
-
-command! {
-    /// Check the status of a transaction. This is usually invoked by a transaction that meets
-    /// another transaction's lock. If the primary lock is expired, it will rollback the primary
-    /// lock. If the primary lock exists but is not expired, it may update the transaction's
-    /// `min_commit_ts`. Returns a [`TxnStatus`](TxnStatus) to represent the status.
-    ///
-    /// This is invoked on a transaction's primary lock. The lock may be generated by either
-    /// [`AcquirePessimisticLock`](Command::AcquirePessimisticLock) or
-    /// [`Prewrite`](Command::Prewrite).
-    CheckTxnStatus:
-        cmd_ty => TxnStatus,
-        display => "kv::command::check_txn_status {} @ {} curr({}, {}) | {:?}", (primary_key, lock_ts, caller_start_ts, current_ts, ctx),
-        content => {
-            /// The primary key of the transaction.
-            primary_key: Key,
-            /// The lock's ts, namely the transaction's start_ts.
-            lock_ts: TimeStamp,
-            /// The start_ts of the transaction that invokes this command.
-            caller_start_ts: TimeStamp,
-            /// The approximate current_ts when the command is invoked.
-            current_ts: TimeStamp,
-            /// Specifies the behavior when neither commit/rollback record nor lock is found. If true,
-            /// rollbacks that transaction; otherwise returns an error.
-            rollback_if_not_exist: bool,
-        }
-}
-
-impl CommandExt for CheckTxnStatus {
-    ctx!();
-    tag!(check_txn_status);
-    ts!(lock_ts);
-    write_bytes!(primary_key);
-    gen_lock!(primary_key);
-}
-
-command! {
-    /// Scan locks from `start_key`, and find all locks whose timestamp is before `max_ts`.
-    ScanLock:
-        cmd_ty => Vec<LockInfo>,
-        display => "kv::scan_lock {:?} {} @ {} | {:?}", (start_key, limit, max_ts, ctx),
-        content => {
-            /// The maximum transaction timestamp to scan.
-            max_ts: TimeStamp,
-            /// The key to start from. (`None` means start from the very beginning.)
-            start_key: Option<Key>,
-            /// The result limit.
-            limit: usize,
-        }
-}
-
-impl CommandExt for ScanLock {
-    ctx!();
-    tag!(scan_lock);
-    ts!(max_ts);
-    command_method!(readonly, bool, true);
-    command_method!(is_sys_cmd, bool, true);
-
-    fn write_bytes(&self) -> usize {
-        0
-    }
-
-    gen_lock!(empty);
-}
-
-command! {
-    /// Resolve locks according to `txn_status`.
-    ///
-    /// During the GC operation, this should be called to clean up stale locks whose timestamp is
-    /// before safe point.
-    ResolveLock:
-        cmd_ty => (),
-        display => "kv::resolve_lock", (),
-        content => {
-            /// Maps lock_ts to commit_ts. If a transaction was rolled back, it is mapped to 0.
-            ///
-            /// For example, let `txn_status` be `{ 100: 101, 102: 0 }`, then it means that the transaction
-            /// whose start_ts is 100 was committed with commit_ts `101`, and the transaction whose
-            /// start_ts is 102 was rolled back. If there are these keys in the db:
-            ///
-            /// * "k1", lock_ts = 100
-            /// * "k2", lock_ts = 102
-            /// * "k3", lock_ts = 104
-            /// * "k4", no lock
-            ///
-            /// Here `"k1"`, `"k2"` and `"k3"` each has a not-yet-committed version, because they have
-            /// locks. After calling resolve_lock, `"k1"` will be committed with commit_ts = 101 and `"k2"`
-            /// will be rolled back.  `"k3"` will not be affected, because its lock_ts is not contained in
-            /// `txn_status`. `"k4"` will not be affected either, because it doesn't have a non-committed
-            /// version.
-            txn_status: HashMap<TimeStamp, TimeStamp>,
-            scan_key: Option<Key>,
-            key_locks: Vec<(Key, Lock)>,
-        }
-}
-
-impl CommandExt for ResolveLock {
-    ctx!();
-    tag!(resolve_lock);
-
-    fn readonly(&self) -> bool {
-        self.key_locks.is_empty()
-    }
-
-    command_method!(is_sys_cmd, bool, true);
-
-    fn write_bytes(&self) -> usize {
-        self.key_locks
-            .iter()
-            .map(|(key, _)| key.as_encoded().len())
-            .sum()
-    }
-
-    gen_lock!(key_locks: multiple(|(key, _)| key));
-}
-
-command! {
-    /// Resolve locks on `resolve_keys` according to `start_ts` and `commit_ts`.
-    ResolveLockLite:
-        cmd_ty => (),
-        display => "kv::resolve_lock_lite", (),
-        content => {
-            /// The transaction timestamp.
-            start_ts: TimeStamp,
-            /// The transaction commit timestamp.
-            commit_ts: TimeStamp,
-            /// The keys to resolve.
-            resolve_keys: Vec<Key>,
-        }
-}
-
-impl CommandExt for ResolveLockLite {
-    ctx!();
-    tag!(resolve_lock_lite);
-    ts!(start_ts);
-    command_method!(is_sys_cmd, bool, true);
-    write_bytes!(resolve_keys: multiple);
-    gen_lock!(resolve_keys: multiple);
-}
-
-command! {
-    /// **Testing functionality:** Latch the given keys for given duration.
-    ///
-    /// This means other write operations that involve these keys will be blocked.
-    Pause:
-        cmd_ty => (),
-        display => "kv::command::pause keys:({}) {} ms | {:?}", (keys.len, duration, ctx),
-        content => {
-            /// The keys to hold latches on.
-            keys: Vec<Key>,
-            /// The amount of time in milliseconds to latch for.
-            duration: u64,
-        }
-}
-
-impl CommandExt for Pause {
-    ctx!();
-    tag!(pause);
-    write_bytes!(keys: multiple);
-    gen_lock!(keys: multiple);
-}
-
-command! {
-    /// Retrieve MVCC information for the given key.
-    MvccByKey:
-        cmd_ty => MvccInfo,
-        display => "kv::command::mvccbykey {:?} | {:?}", (key, ctx),
-        content => {
-            key: Key,
-        }
-}
-
-impl CommandExt for MvccByKey {
-    ctx!();
-    tag!(key_mvcc);
-    command_method!(readonly, bool, true);
-
-    fn write_bytes(&self) -> usize {
-        0
-    }
-
-    gen_lock!(empty);
-}
-
-command! {
-    /// Retrieve MVCC info for the first committed key which `start_ts == ts`.
-    MvccByStartTs:
-        cmd_ty => Option<(Key, MvccInfo)>,
-        display => "kv::command::mvccbystartts {:?} | {:?}", (start_ts, ctx),
-        content => {
-            start_ts: TimeStamp,
-        }
-}
-
-impl CommandExt for MvccByStartTs {
-    ctx!();
-    tag!(start_ts_mvcc);
-    ts!(start_ts);
-    command_method!(readonly, bool, true);
-
-    fn write_bytes(&self) -> usize {
-        0
-    }
-
-    gen_lock!(empty);
 }
 
 impl Command {
