+++ src/storage/txn/commands/prewrite.rs
@@ -1,0 +1,112 @@
+use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::txn::latch::{self, Latches};
+use crate::storage::{Context, Result, TimeStamp};
+use crate::{command, ctx, gen_lock, tag, ts};
+use std::fmt::{self, Debug, Display, Formatter};
+use txn_types::Mutation;
+
+command! {
+    /// The prewrite phase of a transaction. The first phase of 2PC.
+    ///
+    /// This prepares the system to commit the transaction. Later a [`Commit`](Command::Commit)
+    /// or a [`Rollback`](Command::Rollback) should follow.
+    Prewrite:
+        cmd_ty => Vec<Result<()>>,
+        display => "kv::command::prewrite mutations({}) @ {} | {:?}", (mutations.len, start_ts, ctx),
+        content => {
+            /// The set of mutations to apply.
+            mutations: Vec<Mutation>,
+            /// The primary lock. Secondary locks (from `mutations`) will refer to the primary lock.
+            primary: Vec<u8>,
+            /// The transaction timestamp.
+            start_ts: TimeStamp,
+            lock_ttl: u64,
+            skip_constraint_check: bool,
+            /// How many keys this transaction involved.
+            txn_size: u64,
+            min_commit_ts: TimeStamp,
+        }
+}
+
+impl CommandExt for Prewrite {
+    ctx!();
+    tag!(prewrite);
+    ts!(start_ts);
+
+    fn write_bytes(&self) -> usize {
+        let mut bytes = 0;
+        for m in &self.mutations {
+            match *m {
+                Mutation::Put((ref key, ref value)) | Mutation::Insert((ref key, ref value)) => {
+                    bytes += key.as_encoded().len();
+                    bytes += value.len();
+                }
+                Mutation::Delete(ref key) | Mutation::Lock(ref key) => {
+                    bytes += key.as_encoded().len();
+                }
+                Mutation::CheckNotExists(_) => (),
+            }
+        }
+        bytes
+    }
+
+    gen_lock!(mutations: multiple(|x| x.key()));
+}
+
+impl Prewrite {
+    #[cfg(test)]
+    pub fn with_defaults(
+        mutations: Vec<Mutation>,
+        primary: Vec<u8>,
+        start_ts: TimeStamp,
+    ) -> TypedCommand<Vec<Result<()>>> {
+        Prewrite::new(
+            mutations,
+            primary,
+            start_ts,
+            0,
+            false,
+            0,
+            TimeStamp::default(),
+            Context::default(),
+        )
+    }
+
+    #[cfg(test)]
+    pub fn with_lock_ttl(
+        mutations: Vec<Mutation>,
+        primary: Vec<u8>,
+        start_ts: TimeStamp,
+        lock_ttl: u64,
+    ) -> TypedCommand<Vec<Result<()>>> {
+        Prewrite::new(
+            mutations,
+            primary,
+            start_ts,
+            lock_ttl,
+            false,
+            0,
+            TimeStamp::default(),
+            Context::default(),
+        )
+    }
+
+    pub fn with_context(
+        mutations: Vec<Mutation>,
+        primary: Vec<u8>,
+        start_ts: TimeStamp,
+        ctx: Context,
+    ) -> TypedCommand<Vec<Result<()>>> {
+        Prewrite::new(
+            mutations,
+            primary,
+            start_ts,
+            0,
+            false,
+            0,
+            TimeStamp::default(),
+            ctx,
+        )
+    }
+}
