+++ src/storage/txn/commands/txn_heart_beat.rs
@@ -1,0 +1,36 @@
+use crate::storage::metrics::{self, KV_COMMAND_COUNTER_VEC_STATIC};
+use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::txn::latch::{self, Latches};
+use crate::storage::{Context, TimeStamp, TxnStatus};
+use crate::{command, ctx, gen_lock, tag, ts, write_bytes};
+use std::fmt::{self, Debug, Display, Formatter};
+use std::iter;
+use txn_types::Key;
+
+command! {
+    /// Heart beat of a transaction. It enlarges the primary lock's TTL.
+    ///
+    /// This is invoked on a transaction's primary lock. The lock may be generated by either
+    /// [`AcquirePessimisticLock`](Command::AcquirePessimisticLock) or
+    /// [`Prewrite`](Command::Prewrite).
+    TxnHeartBeat:
+        cmd_ty => TxnStatus,
+        display => "kv::command::txn_heart_beat {} @ {} ttl {} | {:?}", (primary_key, start_ts, advise_ttl, ctx),
+        content => {
+            /// The primary key of the transaction.
+            primary_key: Key,
+            /// The transaction's start_ts.
+            start_ts: TimeStamp,
+            /// The new TTL that will be used to update the lock's TTL. If the lock's TTL is already
+            /// greater than `advise_ttl`, nothing will happen.
+            advise_ttl: u64,
+        }
+}
+
+impl CommandExt for TxnHeartBeat {
+    ctx!();
+    tag!(txn_heart_beat);
+    ts!(start_ts);
+    write_bytes!(primary_key);
+    gen_lock!(primary_key);
+}
